Fix Type Marshaller (#1827)
Follow up to https://github.com/microsoft/Power-Fx/pull/1813

+++ src/tests/Microsoft.PowerFx.Interpreter.Tests/ServiceProviderTests.cs
@@ -155,7 +155,7 @@ namespace Microsoft.PowerFx.Interpreter.Tests
             if (expected != null)
             {
                 var runtimeConfig = new RuntimeConfig();
-                runtimeConfig.AddService<User>(new User() { Name = "test", Age = 21 });
+                runtimeConfig.AddService<User>(new User() { Name = "test", Age = 21.0 });
 
                 var res = await check.GetEvaluator().EvalAsync(CancellationToken.None, runtimeConfig).ConfigureAwait(false);
             
@@ -231,7 +231,7 @@ namespace Microsoft.PowerFx.Interpreter.Tests
         {
             public string Name { get; set; }
             
-            public int Age { get; set; }
+            public double Age { get; set; }
         }
     }
 }
